Title: Changed node and launch utilities.
Body: Replaced the use of roslaunch to initiate nodes and launch files
because it generated problems related to threading. New funcionality
uses Python Subprocess to execute nodes and launch files but running
them in a new instanste of XTerm, this way threading is more
manageable and one can launch as many process as desired.

Authored-by: Jose Fajardo <jmfajardod@unal.edu.co>